refactor: move math functions to separate module
- Created new math.py module for mathematical operations
- Moved add, minus, and multiply functions to math.py
- Created dedicated test_math.py for math function tests
- Simplified world.py and test_world.py to focus on world() function
