+++ src/example/math.py
@@ -1,0 +1,30 @@
+def add(a: int, b: int) -> int:
+    """
+    Adds two integers together.
+
+    Args:
+        a: The first integer to add.
+        b: The second integer to add.
+
+    Returns:
+        The sum of the two integers.
+    """
+    return a + b
+
+
+def minus(a: int, b: int) -> int:
+    return a - b
+
+
+def multiply(a: int, b: int) -> int:
+    """
+    Multiplies two integers together.
+
+    Args:
+        a: The first integer to multiply.
+        b: The second integer to multiply.
+
+    Returns:
+        The product of the two integers.
+    """
+    return a * b
